+++ GenumKit/Stencil/Contexts.swift
@@ -144,7 +144,7 @@ extension AssetsCatalogParser {
 
  - `sceneEnumName`: `String`
  - `segueEnumName`: `String`
- - `extraImports`: `Array` of `String`
+ - `modules`: `Array` of `String`
  - `storyboards`: `Array` of:
     - `name`: `String`
     - `initialScene`: `Dictionary` (absent if not specified)
@@ -165,8 +165,7 @@ extension AssetsCatalogParser {
 */
 extension StoryboardParser {
   public func stencilContext(sceneEnumName: String = "StoryboardScene",
-                                           segueEnumName: String = "StoryboardSegue",
-                                           extraImports: [String] = []) -> [String: Any] {
+                                           segueEnumName: String = "StoryboardSegue") -> [String: Any] {
     let storyboards = Set(storyboardsScenes.keys).union(storyboardsSegues.keys).sorted(by: <)
     let storyboardsMap = storyboards.map { (storyboardName: String) -> [String:Any] in
       var sbMap: [String:Any] = ["name": storyboardName]
@@ -174,7 +173,7 @@ extension StoryboardParser {
       if let initialScene = initialScenes[storyboardName] {
         let initial: [String:Any]
         if let customClass = initialScene.customClass {
-          initial = ["customClass": customClass]
+          initial = ["customClass": customClass, "customModule": initialScene.customModule ?? ""]
         } else {
           initial = [
             "baseType": uppercaseFirst(initialScene.tag),
@@ -189,32 +188,35 @@ extension StoryboardParser {
           .sorted(by: {$0.storyboardID < $1.storyboardID})
           .map { (scene: Scene) -> [String:Any] in
             if let customClass = scene.customClass {
-                return ["identifier": scene.storyboardID, "customClass": customClass]
+              return [
+                "identifier": scene.storyboardID,
+                "customClass": customClass,
+                "customModule": scene.customModule ?? ""
+              ]
             } else if scene.tag == "viewController" {
-                return [
-                    "identifier": scene.storyboardID,
-                    "baseType": uppercaseFirst(scene.tag),
-                    "isBaseViewController": scene.tag == "viewController"
-                ]
+              return [
+                "identifier": scene.storyboardID,
+                "baseType": uppercaseFirst(scene.tag),
+                "isBaseViewController": scene.tag == "viewController"
+              ]
             }
             return ["identifier": scene.storyboardID, "baseType": uppercaseFirst(scene.tag)]
         }
       }
       // All Segues
       if let segues = storyboardsSegues[storyboardName] {
-        sbMap["segues"] = segues
-          .sorted(by: {$0.segueID < $1.segueID})
-          .map { (segue: Segue) -> [String:String] in
-            ["identifier": segue.segueID, "customClass": segue.customClass ?? ""]
-        }
+        sbMap["segues"] = segues.sorted(by: {$0.identifier < $1.identifier})
       }
       return sbMap
     }
     return [
       "sceneEnumName": sceneEnumName,
       "segueEnumName": segueEnumName,
-      "extraImports": extraImports,
-      "storyboards": storyboardsMap
+      "storyboards": storyboardsMap,
+      "modules": modules.sorted(),
+
+      // NOTE: This is a deprecated variable
+      "extraImports": modules.sorted()
     ]
   }
 }